requirement.txt
attrs==23.1.0
beautifulsoup4==4.12.2
blinker==1.6.3
certifi==2023.7.22
cffi==1.16.0
click==8.1.7
colorama==0.4.6
dnspython==2.4.2
email-validator==2.0.0.post2
exceptiongroup==1.1.3
flask==3.0.0
flask-wtf==1.2.1
h11==0.14.0
idna==3.4
importlib-metadata==6.8.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
mysql-connector==2.2.9
numpy==1.26.0
outcome==1.2.0
pandas==2.1.1
pycparser==2.21
PySocks==1.7.1
python-dateutil==2.8.2
pytz==2023.3.post1
selenium==4.14.0
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.5
trio==0.22.2
trio-websocket==0.11.1
tzdata==2023.3
urllib3==2.0.6
werkzeug==3.0.0
wsproto==1.2.0
wtforms==3.1.0
zipp==3.17.0
